+++ packages/webview_flutter/webview_flutter/test/webview_flutter_test.mocks.dart
@@ -87,6 +87,11 @@ class MockWebViewPlatformController extends _i1.Mock
           returnValue: Future<void>.value(),
           returnValueForMissingStub: Future<void>.value()) as _i9.Future<void>);
   @override
+  _i9.Future<void> loadFlutterAsset(String? key) =>
+      (super.noSuchMethod(Invocation.method(#loadFlutterAsset, [key]),
+          returnValue: Future<void>.value(),
+          returnValueForMissingStub: Future<void>.value()) as _i9.Future<void>);
+  @override
   _i9.Future<void> loadHtmlString(String? html, {String? baseUrl}) =>
       (super.noSuchMethod(
           Invocation.method(#loadHtmlString, [html], {#baseUrl: baseUrl}),