+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview.dart
@@ -3,10 +3,12 @@
 // found in the LICENSE file.
 
 import 'dart:typed_data';
+import 'dart:ui';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart' show AndroidViewSurface;
 
+import 'android_webview.pigeon.dart';
 import 'android_webview_api_impls.dart';
 
 // TODO(bparrishMines): This can be removed once pigeon supports null values: https://github.com/flutter/flutter/issues/59118
@@ -356,6 +358,11 @@ class WebView {
     );
     WebChromeClient.api.createFromInstance(client, _currentWebViewClient!);
     return api.setWebChromeClientFromInstance(this, client);
+  }
+
+  /// Sets the background color of this WebView.
+  Future<void> setBackgroundColor(Color color) {
+    return api.setBackgroundColorFromInstance(this, color.value);
   }
 
   /// Releases all resources used by the [WebView].
@@ -771,4 +778,24 @@ class WebResourceError {
 
   /// Describes the error.
   final String description;
+}
+
+/// Manages Flutter assets that are part of Android's app bundle.
+class FlutterAssetManager {
+  /// Constructs the [FlutterAssetManager].
+  const FlutterAssetManager();
+
+  /// Pigeon Host Api implementation for [FlutterAssetManager].
+  @visibleForTesting
+  static FlutterAssetManagerHostApi api = FlutterAssetManagerHostApi();
+
+  /// Lists all assets at the given path.
+  ///
+  /// The assets are returned as a `List<String>`. The `List<String>` only
+  /// contains files which are direct childs
+  Future<List<String?>> list(String path) => api.list(path);
+
+  /// Gets the relative file path to the Flutter asset with the given name.
+  Future<String> getAssetFilePathByName(String name) =>
+      api.getAssetFilePathByName(name);
 }